Start to split StandardFunctionHandler into different classes
